Blog/requirements.txt
agate==1.6.1
agate-dbf==0.2.2
agate-excel==0.2.3
agate-sql==0.5.5
aiohttp==3.7.3
ansi2html==1.5.2
arabic-reshaper==2.1.3
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.1
csvkit==1.0.4
cycler==0.10.0
dbfread==2.0.7
dj-database-url==0.5.0
Django==2.2.13
django-cors-headers==3.0.2
django-filter==2.1.0
django-heroku==0.3.1
django-queryset-csv==1.0.2
django-restframework==0.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.10.3
djangorestframework-jsonp==1.0.2
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
future==0.18.2
gitdb==4.0.5
gitdb2==4.0.2
GitPython==3.0.6
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.8
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
leather==0.3.3
lxml==4.5.0
Markdown==3.1.1
matplotlib==3.3.4
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.1
openpyxl==3.0.6
packaging==20.9
pandas==1.2.2
parsedatetime==2.6
Pillow==8.1.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-slugify==4.0.1
pytimeparse==1.1.8
pytz==2020.1
reportlab==3.5.67
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.13.0
slackclient==2.1.0
smmap==3.0.5
soupsieve==2.2
SQLAlchemy==1.3.23
sqlparse==0.4.0
tabulate==0.8.2
text-unidecode==1.3
tldextract==2.2.0
truffleHog==2.2.1
truffleHogRegexes==0.0.7
typing-extensions==3.7.4.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.11.2
xhtml2pdf==0.2.5
xlrd==2.0.1
yarl==1.6.3
